\name{normalHypothesesPlotBoxCox}
\alias{normalHypothesesPlot}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Rcmdr normalHypothesesPlot menu.}
\description{
  Rcmdr normalHypothesesPlot menu
}
\usage{
normalHypothesesPlot()
}
%- maybe also 'usage' for other objects documented here.
\author{Richard M. Heiberger <rmh@temple.edu>.}
\seealso{\code{\link{norm.curve}}}
\keyword{hplot}
