\name{totPercents2.HH}
\alias{totPercents2.HH}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Calculate Total Percents for a Two Way Table}
\description{
Divide every number in the table by the grand total and multiply by 100.
}
\usage{
totPercents2.HH(tab, digits = 1)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{tab}{Two way table or data.frame or matrix.}
  \item{digits}{number of digits passed to the \code{round} function.}
}
\value{
Divide every number in the table by the grand total and multiply by 100.
}
\author{ Richard M. Heiberger <rmh@temple.edu> }
\keyword{ array }
